Bug 14967 MapRFS counters value show negative values
Looks like large number of read operations are causing a overflow. Changing the data type for readOps, largeReadOps, writeOps to long from int.
The statistics for file system operations are collected here.
http://10.10.30.177:8080/source/xref/hadoop-common-2.3-mapr/hadoop-mapreduce-project/hadoop-mapreduce-client/hadoop-mapreduce-client-core/src/main/java/org/apache/hadoop/mapred/Task.java#931

+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/fs/FileSystem.java
@@ -2844,9 +2844,9 @@ public abstract class FileSystem extends Configured implements Closeable {
     public static class StatisticsData {
       volatile long bytesRead;
       volatile long bytesWritten;
-      volatile int readOps;
-      volatile int largeReadOps;
-      volatile int writeOps;
+      volatile long readOps;
+      volatile long largeReadOps;
+      volatile long writeOps;
       /**
        * Stores a weak reference to the thread owning this StatisticsData.
        * This allows us to remove StatisticsData objects that pertain to
